+++ lib/emque/producing/configuration.rb
@@ -9,18 +9,20 @@ module Emque
       attr_accessor :log_publish_message
       attr_accessor :publish_messages
       attr_reader :rabbitmq_options
+      attr_reader :google_cloud_pubsub_options
       attr_accessor :ignored_exceptions
       attr_reader :middleware
 
       def initialize
         @app_name = ""
-        @publishing_adapter = :rabbitmq
+        @publishing_adapter = [:rabbitmq]
         @error_handlers = []
         @log_publish_message = false
         @publish_messages = true
         @rabbitmq_options = {
           :url => "amqp://guest:guest@localhost:5672"
         }
+        @google_cloud_pubsub_options = {}
         @ignored_exceptions = [Emque::Producing::Message::MessagesNotSentError]
         @middleware = []
       end